+++ alpha/apps/kaltura/lib/batch2/kFlowHelper.php
@@ -25,6 +25,8 @@ class kFlowHelper
 
 	const DAYS = 'days';
 	const HOURS = 'hours';
+
+	const MAX_USER_ENTRIES_TO_RECYCLE = 50;
 	
 	
 	/**
@@ -3033,6 +3035,16 @@ class kFlowHelper
 		$userEntryCriteria->add(UserEntryPeer::ENTRY_ID, $entry->getId());
 		$userEntryCriteria->add(UserEntryPeer::STATUS, array($oldUserEntryStatus), Criteria::IN);
 		UserEntryPeer::setUseCriteriaFilter(false);
+
+		$userEntriesCount = UserEntryPeer::doCount($userEntryCriteria);
+		KalturaLog::info("Found $userEntriesCount user entries for entry id [{$entry->getId()}] with status [$oldUserEntryStatus] to update to status [$newUserEntryStatus]");
+		$groupLimit = kConf::get('recycle_bin_userentry_threshold', kConfMapNames::LOCAL_SETTINGS, self::MAX_USER_ENTRIES_TO_RECYCLE);
+		if ($groupLimit < $userEntriesCount)
+		{
+			kJobsManager::addUpdateUserEntriesJob($entry->getPartnerId(), $entry->getId(), $oldUserEntryStatus, $newUserEntryStatus);
+			return;
+		}
+
 		$userEntriesList = UserEntryPeer::doSelect($userEntryCriteria);
 		UserEntryPeer::setUseCriteriaFilter(true);
 		if (!$userEntriesList)